Ajustes de ToolTip en web OK
